% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{BoxModification}
\alias{BoxModification}
\title{Plot boxplot for each modification within each brain region}
\usage{
BoxModification(MergeModificationIntensityFile, SampleIndex)
}
\arguments{
\item{SampleIndex}{provide the sample prefix to set brain region names}

\item{MergeModificationIntensityName}{A directory stores calculated files generated by MergeModificationIntensity()}
}
\description{
Plot boxplot for each modification within each brain region
}
\examples{

BoxModification(MergeModificationIntensityFile = "2022-03-04 mettl3-KO-loxP/1_Negative-merged-NormalizedSIndividual.csv",
                SampleIndex = c("Neg-mettl3-loxP1-","Neg-mettl3-loxP2-","Neg-mettl3-loxP3-",
                                "Neg-mettl3-KO1-","Neg-mettl3-KO2-","Neg-mettl3-KO3-"))
}
