% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{CalculateModificationIntensity}
\alias{CalculateModificationIntensity}
\title{Calculate signal intensity of each modification from MS results}
\usage{
CalculateModificationIntensity(MSFilePath, ModificationReferenceFile)
}
\arguments{
\item{MSFilePath}{A directory containing excel xlsx files that records m/z and corresponding signal intensity}

\item{ModificationReferenceFile}{A user defined reference modification list and m/z ranges}
}
\description{
Calculate signal intensity of each modification from MS results
}
\examples{

CalculateModificationIntensity(MSFile = "2022-01-10 Negative-WT1-14 brain regions-03.xlsx",
                               ModificationReferenceFile = "Modification_reference.csv")
}
