% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{GenrateExcelforMSdata}
\alias{GenrateExcelforMSdata}
\title{Generate excel tables for store MS-imaging data for each region, each section and each slide}
\usage{
GenrateExcelforMSdata(BrainRegionNames, SamplesNames, SectionNames)
}
\arguments{
\item{BrainRegionNames}{Brain region names selected in the MS-imaging}

\item{SamplesNames}{Sample names for each section}

\item{SectionNames}{Number for each section}
}
\description{
Generate excel tables for store MS-imaging data for each region, each section and each slide
}
\examples{

GenrateExcelforMSdata(BrainRegionNames = c("cerebellum","pons&medulla","midbrain","hippocampus","thalamus",
                                           "hypothalamus","fornix","caudate putamen","basal forebrain","ventral striatum",
                                           "anterior olfactory","olfactory bulbs","cortex","corpus callosum"),
                     SamplesNames = c("WT4","WT5","WT6","5-FAD4","5-FAD5","5-FAD6"),
                     SectionNames = c("01","02","03","04"))
}
