% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{MergeModificationIntensity}
\alias{MergeModificationIntensity}
\title{Merged csv files returned by the CalculateModificationIntensity() function}
\usage{
MergeModificationIntensity(
  CSVPath = "",
  PickBrainRegionNames = c(),
  run_RunMetaboAnalystR = T
)
}
\arguments{
\item{CSVPath}{A directory store calculated files generated}

\item{PickBrainRegionNames}{Specify Brain region names for output}

\item{run_RunMetaboAnalystR}{Whether run MetaboAnalystR analysis in the meanwhile}
}
\description{
Merged csv files returned by the CalculateModificationIntensity() function
}
\examples{

MergeModificationIntensity(CSVPath = "./Negative",
                           PickBrainRegionNames = c("hippocampus","cerebellum"))
}
