% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{MergeTechnicalReplicates}
\alias{MergeTechnicalReplicates}
\title{Merge technical replicates based on section suffix}
\usage{
MergeTechnicalReplicates(MergeModificationIntensityFile)
}
\arguments{
\item{MergeModificationIntensityFile}{The merged modification intensity file generated by MergeModificationIntensity()}
}
\description{
Merge technical replicates based on section suffix
}
\examples{

MergeTechnicalReplicates(MergeModificationIntensityFile = "Positive-merged.csv")
}
