% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{MergeTR}
\alias{MergeTR}
\title{Merge technical replicates based on section suffix}
\usage{
MergeTR(PlainModificationIntensityName = "")
}
\arguments{
\item{PlainModificationIntensityName}{A modified version of MergeModificationIntensityName}
}
\description{
Merge technical replicates based on section suffix
}
\examples{

MergeTR(PlainModificationIntensityName = "Positive-merged.csv")
}
