% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{NormalizationIntensityPercentage}
\alias{NormalizationIntensityPercentage}
\title{Normalize each modification to the total amount of its corresponding nucleotide base}
\usage{
NormalizationIntensityPercentage(
  MergeModificationIntensityFile,
  ModificationReferenceFile
)
}
\arguments{
\item{MergeModificationIntensityFile}{The merged modification intensity file generated by MergeModificationIntensity()}

\item{ModificationReferenceFile}{A user defined reference modification list and m/s ranges}
}
\description{
Normalize each modification to the total amount of its corresponding nucleotide base
}
\examples{

NormalizationIntensityPercentage(MergeModificationIntensityFile = "2022-03-04 mettl3-KO-loxP/1_Negative-merged.csv",
                                 ModificationReferenceFile = "2022-03-04 mettl3-KO-loxP/Re.csv")
}
