% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{NormalizationIntensitySections}
\alias{NormalizationIntensitySections}
\title{Normalize signal in each region by dividing total signals in all region in the same section}
\usage{
NormalizationIntensitySections(
  MergeModificationIntensityFile,
  SampleIndex,
  SectionIndex
)
}
\arguments{
\item{MergeModificationIntensityFile}{The merged modification intensity file generated by MergeModificationIntensity()}

\item{SampleIndex}{Specify Sample index for group samples}

\item{SectionIndex}{Specify Section index for group sections}
}
\description{
Normalize signal in each region by dividing total signals in all region in the same section
}
\examples{

NormalizationIntensity(MergeModificationIntensityFile = "",
                       SampleIndex = c("FAD1","FAD2","FAD3","WT1","WT2","WT3"),
                       SectionIndex = c("-01","-02","-03","-04"))
}
