% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{NormalizationIntensitySlides}
\alias{NormalizationIntensitySlides}
\title{Normalize signal in each region by dividing total signals in all region and all sections in the same slide}
\usage{
NormalizationIntensitySlides(MergeModificationIntensityFile, SectionIndex)
}
\arguments{
\item{MergeModificationIntensityFile}{The merged modification intensity file generated by MergeModificationIntensity()}

\item{SectionIndex}{Specify Section index for group sections}
}
\description{
Normalize signal in each region by dividing total signals in all region and all sections in the same slide
}
\examples{

NormalizationIntensity(MergeModificationIntensityFile = "",
                       SectionIndex = c("-01","-02","-03","-04"))
}
