% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{PickBrainRegion}
\alias{PickBrainRegion}
\title{This is a standalone script for picking certain regions and features}
\usage{
PickBrainRegion(
  MergeModificationIntensityName = "",
  PickBrainRegionNames = c(),
  run_RunMetaboAnalystR = T,
  FeatureList = c()
)
}
\arguments{
\item{MergeModificationIntensityName}{A directory stores calculated files generated by MergeModificationIntensity()}

\item{PickBrainRegionNames}{Specify Brain region name(s) for output}

\item{run_RunMetaboAnalystR}{Whether run MetaboAnalystR analysis in the meanwhile}
}
\description{
This is a standalone script for picking certain regions and features
}
\examples{

PickBrainRegion(MergeModificationIntensityName = "Positive-merged.csv",
                PickBrainRegionNames = "cerebellum",
                run_RunMetaboAnalystR = T,
                FeatureList = c("m6Am","ac4C","m22G","hm5CTP","m6dA","m5dC",
                                "ca5dC","m5dCTP","m6dATP","f5dCTP","m5CMP","m6AMP"))
}
