% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{RunMetaboAnalystR}
\alias{RunMetaboAnalystR}
\title{Run MetaboAnalystR codes}
\usage{
RunMetaboAnalystR(pktablePath, rowNormMet = c("SumNorm", "NULL")[2])
}
\arguments{
\item{pktablePath}{A standard MetaboAnalystR input file}

\item{rowNormMet}{method for column normalization, recommend using NULL for data has internal controls}
}
\description{
Run MetaboAnalystR codes
}
\examples{

RunMetaboAnalystR(pktablePath = "refinedDF_5depots_abbrev_adi_peri+adi_ibat.csv",
                 rowNormMet = c("SumNorm","NULL")[2])
}
