% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{T_Anavo_plotting}
\alias{T_Anavo_plotting}
\title{Draw costume scatter plot for pair-wise T.test or multi-group anova result}
\usage{
T_Anavo_plotting(
  PlotFile,
  SigIndex = c("p_value", "fdr")[1],
  ThresholdSig = 0.005,
  TopShown = 10
)
}
\arguments{
\item{PlotFile}{The file returned by Volcano.Anal()}

\item{SigIndex}{select using raw p-value (p_value) or adjusted p-value (fdr)}

\item{ThresholdSig}{Threshold for significance}

\item{TopShown}{Top significant items for labeling}
}
\description{
Draw costume scatter plot for pair-wise T.test or multi-group anova result
}
\examples{

T_Anavo_plotting(PlotFile = "refinedDF_5depots_abbrev_anova_posthoc.csv",
                 SigIndex = c("p_value","fdr")[1],
                 ThresholdSig = 0.005,TopShown = 10)
}
