% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{volcano_plotting}
\alias{volcano_plotting}
\title{Draw modified Volcano Plot using ggplot}
\usage{
volcano_plotting(
  PlotFile,
  ThresholdFC = 1.5,
  ThresholdSig = 0.05,
  TopUpDownShown = c(10, 10)
)
}
\arguments{
\item{PlotFile}{The file returned by Volcano.Anal() in MetaboAnalystR}

\item{ThresholdFC}{Threshold for fold-change}

\item{ThresholdSig}{Threshold for significance}

\item{TopUpDownShown}{Select top N up, down significant features for shown}
}
\description{
Draw modified Volcano Plot using ggplot
}
\examples{

volcano_plotting(PlotFile = "refinedDF_5depots_abbrev_adi_epi+adi_ibat_volcano.csv",
                 ThresholdFC = 1.5,
                 ThresholdSig = 0.05,
                 TopUpDownShown = c(10,10))
}
