% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Modifications.R
\name{MergeModificationIntensity}
\alias{MergeModificationIntensity}
\title{Merged csv files returned by the CalculateModificationIntensity() function}
\usage{
MergeModificationIntensity(CSVPath, run_RunMetaboAnalystR = F)
}
\arguments{
\item{CSVPath}{A directory store calculated files generated}

\item{run_RunMetaboAnalystR}{Whether run MetaboAnalystR analysis for the merged dataset}
}
\description{
Merged csv files returned by the CalculateModificationIntensity() function
}
\examples{

MergeModificationIntensity(CSVPath = "./Negative",
                           PickBrainRegionNames = c("hippocampus","cerebellum"))
}
